requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
bitarray==3.0.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
crc16==0.1.1
crc32c==2.7.1
cycler==0.12.1
Django==5.0.6
environs==11.0.0
fonttools==4.54.1
frozenlist==1.4.1
graphql-core==3.2.5
graphql_query==1.0.3
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.7
magic-filter==1.0.12
MarkupSafe==3.0.2
marshmallow==3.21.3
matplotlib==3.9.2
multidict==6.0.5
numpy==2.1.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydantic==1.10.19
pydantic_core==2.20.1
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytonlib==0.0.63
requests==2.32.3
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.0
ton==0.26
tonsdk==1.0.15
TonTools==2.1.2
tvm-valuetypes==0.0.12
typing_extensions==4.12.2
urllib3==2.2.1
yarl==1.9.4
